+++ src/helpers.ts
@@ -65,9 +65,13 @@ export function transformKeys<T>(data: T, transform: (key: string) => string): T
 
 
 // Add 1 to the number at counts[key], or set it to 1 if it doesn't exist
-export function incrementCount(counts: StringNumberDict, key?: CountKey | null): StringNumberDict {
+export function incrementCount(
+    counts: StringNumberDict,
+    key?: CountKey | null,
+    increment: number = 1
+): StringNumberDict {
     key = key ?? "unknown";
-    counts[key] = (counts[key] || 0) + 1;
+    counts[key] = (counts[key] || 0) + increment;
     return counts;
 };